Fix order/discount display bugs, unify all combo styling, hide out-of-stock products
- Show live discount-code usage count instead of an always-zero column, and
  break out promotion/coupon discounts on the admin order detail page and
  order confirmation emails (previously only the final total was shown).
- Persist the applied coupon across page reloads (cart was losing it).
- Add a shared app-ui-select component so every fixed-option dropdown
  (admin filters/forms, order and inquiry status, public catalog filters)
  renders with the same custom panel as the existing search combos, instead
  of the browser's native <select> styling; also fixes a hover-state CSS
  conflict with the global primary-button style.
- Hide products with zero stock from the public catalog, related-products
  widget, and sitemap, and 404 their detail page instead of showing an
  empty/unbuyable listing.

+++ backend/src/main/java/uy/washop/discount/api/dto/DiscountCodeResponse.java
@@ -12,7 +12,7 @@ public record DiscountCodeResponse(
         DiscountType discountType,
         BigDecimal discountValue,
         Integer maxUses,
-        int usedCount,
+        long usedCount,
         Instant startsAt,
         Instant endsAt,
         Instant createdAt,
